@@ -1,8 +1,8 @@
 package de.tuberlin.mcc.simra.app.util;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+import androidx.appcompat.app.AppCompatActivity;
 import de.tuberlin.mcc.simra.app.R;
 
 public class HelpActivity extends AppCompatActivity {
